e_com/requirements.txt
asgiref==3.6.0
Django>=3
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
gunicorn==20.0.4
